+++ api_golden_test.go
@@ -289,6 +289,8 @@ func (c *Client) StartObservation(
 
 func (c *Client) WithBaggagePropagation(ctx context.Context) context.Context
 
+func (c *Client) WithContentCapture(ctx context.Context, enabled bool) context.Context
+
 func (c *Client) WithSampleRate(ctx context.Context, fraction float64) context.Context
 
 func (c *Client) WithTraceAttributes(ctx context.Context, values TraceAttributes) context.Context